+++ tests/functional/AbstractBigBlueButtonFunctionalTest.php
@@ -20,7 +20,6 @@
 namespace BigBlueButton\Tests\Functional;
 
 use BigBlueButton\BigBlueButton;
-use BigBlueButton\Core\ApiMethod;
 use BigBlueButton\Http\Transport\TransportInterface;
 use BigBlueButton\Parameters\DeleteRecordingsParameters;
 use BigBlueButton\Parameters\EndMeetingParameters;
@@ -98,17 +97,6 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
     }
 
     /* Create Meeting */
-
-    /**
-     * Test create meeting URL.
-     */
-    public function testCreateMeetingUrl(): void
-    {
-        $params = $this->generateCreateParams();
-        $url = $this->bbb->getCreateMeetingUrl($this->getCreateMock($params));
-
-        $this->assertUrlContainsAllRequestParameters($url, $params);
-    }
 
     /**
      * Test create meeting.
@@ -198,19 +186,6 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
 
     /* Join Meeting */
 
-    /**
-     * Test create join meeting URL.
-     */
-    public function testCreateJoinMeetingUrl(): void
-    {
-        $joinMeetingParams = $this->generateJoinMeetingParams();
-        $joinMeetingMock = $this->getJoinMeetingMock($joinMeetingParams);
-
-        $url = $this->bbb->getJoinMeetingURL($joinMeetingMock);
-
-        $this->assertUrlContainsAllRequestParameters($url, $joinMeetingParams);
-    }
-
     public function testJoinMeeting()
     {
         $params = $this->getCreateMock($this->generateCreateParams());
@@ -238,17 +213,6 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
     }
 
     /* End Meeting */
-
-    /**
-     * Test generate end meeting URL.
-     */
-    public function testCreateEndMeetingUrl(): void
-    {
-        $params = $this->generateEndMeetingParams();
-        $url = $this->bbb->getEndMeetingURL($this->getEndMeetingMock($params));
-
-        $this->assertUrlContainsAllRequestParameters($url, $params);
-    }
 
     public function testEndMeeting()
     {
@@ -280,12 +244,6 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
 
     /* Get Meetings */
 
-    public function testGetMeetingsUrl(): void
-    {
-        $url = $this->bbb->getMeetingsUrl();
-        $this->assertStringContainsString(ApiMethod::GET_MEETINGS, $url);
-    }
-
     public function testGetMeetings()
     {
         $meeting = $this->createRealMeeting($this->bbb);
@@ -299,17 +257,6 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
 
     /* Get meeting info */
 
-    public function testGetMeetingInfoUrl(): void
-    {
-        $meeting = $this->createRealMeeting($this->bbb);
-
-        $url = $this->bbb->getMeetingInfoUrl(new GetMeetingInfoParameters($meeting->getMeetingId()));
-        $this->assertStringContainsString('='.rawurlencode($meeting->getMeetingId()), $url);
-
-        // Cleanup
-        $this->bbb->endMeeting(new EndMeetingParameters($meeting->getMeetingId(), $meeting->getModeratorPassword()));
-    }
-
     public function testGetMeetingInfo()
     {
         $meeting = $this->createRealMeeting($this->bbb);
@@ -322,23 +269,11 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
         $this->bbb->endMeeting(new EndMeetingParameters($meeting->getMeetingId(), $meeting->getModeratorPassword()));
     }
 
-    public function testGetRecordingsUrl(): void
-    {
-        $url = $this->bbb->getRecordingsUrl(new GetRecordingsParameters());
-        $this->assertStringContainsString(ApiMethod::GET_RECORDINGS, $url);
-    }
-
     public function testGetRecordings()
     {
         $result = $this->bbb->getRecordings(new GetRecordingsParameters());
         $this->assertEquals('SUCCESS', $result->getReturnCode());
         $this->assertTrue($result->success());
-    }
-
-    public function testPublishRecordingsUrl(): void
-    {
-        $url = $this->bbb->getPublishRecordingsUrl(new PublishRecordingsParameters($this->faker->sha1, true));
-        $this->assertStringContainsString(ApiMethod::PUBLISH_RECORDINGS, $url);
     }
 
     public function testPublishRecordings()
@@ -348,29 +283,11 @@ abstract class AbstractBigBlueButtonFunctionalTest extends TestCase
         $this->assertTrue($result->failed());
     }
 
-    public function testDeleteRecordingsUrl(): void
-    {
-        $url = $this->bbb->getDeleteRecordingsUrl(new DeleteRecordingsParameters($this->faker->sha1));
-        $this->assertStringContainsString(ApiMethod::DELETE_RECORDINGS, $url);
-    }
-
     public function testDeleteRecordings()
     {
         $result = $this->bbb->deleteRecordings(new DeleteRecordingsParameters('non-existing-id-'.$this->faker->sha1));
         $this->assertEquals('FAILED', $result->getReturnCode());
         $this->assertTrue($result->failed());
-    }
-
-    public function testUpdateRecordingsUrl(): void
-    {
-        $params = $this->generateUpdateRecordingsParams();
-        $url = $this->bbb->getUpdateRecordingsUrl($this->getUpdateRecordingsParamsMock($params));
-        foreach ($params as $key => $value) {
-            if (\is_bool($value)) {
-                $value = $value ? 'true' : 'false';
-            }
-            $this->assertStringContainsString('='.rawurlencode($value), $url);
-        }
     }
 
     public function testUpdateRecordings()
